requirements.txt
pymupdf
pydantic
google-generativeai
python-dotenv
pinecone
openai